+++ CHANGELOG.md
@@ -11,7 +11,7 @@ but cannot always guarantee backwards compatibility. Changes that may **break co
 
 **Improved**
 
-- New Feature: By setting `user_reversible_instance_norm=True`, the reversible [instance normalization](https://github.com/ts-kim/RevIN) is be applied to the target series. This was already available for `PLForecastingModule` and is now also available for `RegressionModel`. By [Jules Authier](https://github.com/julesauthier).
+- New Feature: By setting `user_reversible_instance_norm=True`, the reversible [instance normalization](https://github.com/ts-kim/RevIN) is applied to the target series. This was already available for `PLForecastingModule` and is now also available for `RegressionModel`. By [Jules Authier](https://github.com/julesauthier).
 - New model: `StatsForecastAutoTBATS`. This model offers the [AutoTBATS](https://nixtlaverse.nixtla.io/statsforecast/src/core/models.html#autotbats) model from Nixtla's `statsforecasts` library. [#2611](https://github.com/unit8co/darts/pull/2611) by [He Weilin](https://github.com/cnhwl).
 - Added the `title` attribute to `TimeSeries.plot()`. This allows to set a title for the plot. [#2639](https://github.com/unit8co/darts/pull/2639) by [Jonathan Koch](https://github.com/jonathankoch99).
 - Added parameter `component_wise` to `show_anomalies()` to separately plot each component in multivariate series. [#2544](https://github.com/unit8co/darts/pull/2544) by [He Weilin](https://github.com/cnhwl).